updated aging

+++ paging.c
@@ -76,10 +76,10 @@ int aging()
 {
   int i, min = 0;
   for(i = 0; i < mem_size; i++){
-    if(page_table[mem[i]].counter < page_table[min].counter){
+    if((int)page_table[mem[i]].counter < page_table[min].counter){
       min = i;
     }
-    else if(page_table[mem[i]].counter == page_table[min].counter){
+    else if((int)page_table[mem[i]].counter == page_table[min].counter){
       return min;
     }
   }
